pyproject.toml
[project]
name = "px-plus"
version = "0.1.0"
description = "FastAPI + RFS Framework Hello World Project"
authors = [
    {name = "Sangbong Moon", email = "sangbong@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "rfs-framework>=4.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "spacy>=3.7.2",
    "python-pptx>=0.6.21",
    "python-docx>=0.8.11",
    "openpyxl>=3.1.2",
    "pypdf>=3.9.0",
    "google-genai>=1.33.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "mypy>=1.3.0",
    "isort>=5.12.0",
    "httpx>=0.25.0",  # FastAPI 테스트용
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # 기본 제외 디렉토리
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-report=html"
